# ssr_immemory/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/request-paints/{request_id}/{paint_id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет данные краски в заявке (площадь, слои, количество)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "request-paints"
                ],
                "summary": "Изменить краску в заявке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "request_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID краски",
                        "name": "paint_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные краски в заявке",
                        "name": "requestPaint",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apitypes.RequestsPaintJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apitypes.RequestsPaintJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет краску из указанной заявки",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "request-paints"
                ],
                "summary": "Удалить краску из заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "request_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID краски",
                        "name": "paint_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apitypes.PaintRequestJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apitypes.PaintRequestJSON": {
            "type": "object",
            "properties": {
                "creator_login": {
                    "type": "string"
                },
                "dateFinish": {
                    "type": "string"
                },
                "dateForm": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "min_layers": {
                    "type": "integer"
                },
                "moderator_login": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "apitypes.RequestsPaintJSON": {
            "type": "object",
            "properties": {
                "area": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "layers": {
                    "type": "integer"
                },
                "paintID": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "requestID": {
                    "type": "integer"
                }
            }
        },
...
    }
}